Backend/python/Scheduler_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pandas==2.2.3
python-multipart==0.0.12
supabase==2.6.0
python-dotenv==1.0.1